requirements/base.txt
asgiref==3.2.3
Django==3.0.4
pytz==2019.3
sqlparse==0.3.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
coverage==5.0.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
